+++ workspace/lint-staged.config.js
@@ -5,7 +5,12 @@
 const config = {
     // REF: https://github.com/sudo-suhas/lint-staged-django-react-demo
     "*": "cspell lint --no-must-find-files",
-    "*.{js,ts,jsx,tsx,vue,svelte,md,json,toml,yaml}": "eslint --fix",
+    "*.{js,ts}": "eslint --fix",
+    "*.{css,scss,less,pcss,postcss}": "eslint --fix",
+    "*.{jsx,tsx,vue,svelte}": "eslint --fix",
+    "*.{md,htm,html}": "eslint --fix",
+    "*.{xml,json,jsonc,json5,yml,yaml,toml}": "eslint --fix",
+    "*.{svg,xbel}": "eslint --fix",
 };
 
 export default config;